Bug fix: Weekly new tests and "cases" (ie. positive test results) count should be aggregated from their daily counts rather than derived as their daily mean over a week.
Seems like a huge bug, but efectively it was meaningless for my
purposes. new_cases_smoothed and new_tests_smoothed are only used to
calculate people_vaccinated_percent, which is then drawn. As mean
and sum of a series are linearly propotional to each other, equation
`new_cases_smoothed / new_tests_smoothed * 100` yields an almost
identical result for both cases. Which is why I haven't ever noticed
this bug on charts. It must have been due to a copy/paste from another
line.